fix cataas request status

+++ user-service/user/endpoints/user.py
@@ -65,7 +65,7 @@ def create_user(request):
     # add default picture
     try:
         cat_request = requests.get("https://cataas.com/cat?type=square&position=center")
-        if (cat_request.status_code / 100) % 10 == 2:
+        if int((cat_request.status_code / 100) % 10) == 2:
             with open(f"{settings.PICTURES_DST}/{new_user.id}.png", "wb+") as f:
                 f.write(cat_request.content)
         else:
